requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
redis==5.0.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.4
tavily-python==0.3.3
ollama==0.3.3
pgvector==0.2.5
numpy==1.26.4
langgraph==0.2.55
langchain-core==0.3.28
pypdf==4.2.0
python-multipart==0.0.9
streamlit==1.35.0
requests==2.32.3